examples/python-telegram-bot/requirements.txt
python-telegram-bot~=13.11
